refactor: サービスのパッケージ階層を整理。
